# packages/remote-server/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VCKit API"
  },
  "servers": [],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/didManagerGetProviders": {
      "post": {
        "operationId": "didManagerGetProviders",
        "description": "Returns a list of available identifier providers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a list of available identifier providers",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/didManagerFind": {
      "post": {
        "operationId": "didManagerFind",
        "description": "Returns a list of managed identifiers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IDIDManagerFindArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a list of managed identifiers",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IIdentifier"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/didManagerGet": {
      "post": {
        "operationId": "didManagerGet",
        "description": "Returns a specific identifier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IDIDManagerGetArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a specific identifier",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/IIdentifier"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/didManagerCreate": {
      "post": {
        "operationId": "didManagerCreate",
        "description": "Creates and returns a new identifier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IDIDManagerCreateArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Creates and returns a new identifier",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/IIdentifier"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/didManagerDelete": {
      "post": {
        "operationId": "didManagerDelete",
        "description": "Deletes identifier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IDIDManagerDeleteArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletes identifier",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/dataStoreGetVerifiableCredential": {
      "post": {
        "operationId": "dataStoreGetVerifiableCredential",
        "description": "Gets verifiable credential from the data store",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IDataStoreGetVerifiableCredentialArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gets verifiable credential from the data store",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiableCredential"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/dataStoreORMGetIdentifiersCount": {
      "post": {
        "operationId": "dataStoreORMGetIdentifiersCount",
        "description": "Tries to obtain a count of ",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindArgs-TIdentifiersColumns"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tries to obtain a count of ",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/dataStoreORMGetVerifiableCredentials": {
      "post": {
        "operationId": "dataStoreORMGetVerifiableCredentials",
        "description": "Returns a list of ",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindArgs-TCredentialColumns"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a list of ",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniqueVerifiableCredential"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/dataStoreORMGetVerifiableCredentialsCount": {
      "post": {
        "operationId": "dataStoreORMGetVerifiableCredentialsCount",
        "description": "Returns a count of ",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindArgs-TCredentialColumns"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a count of ",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/renderCredential": {
      "post": {
        "operationId": "renderCredential",
        "description": "Renders a verifiable credential.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IRenderCredentialArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Renders a verifiable credential.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/IRenderResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/fetchEncryptedDataByCredentialHash": {
      "post": {
        "operationId": "fetchEncryptedDataByCredentialHash",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IFetchEncryptedDataByCredentialHashArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/IFetchEncryptedDataByCredentialHashResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/issueRevocationStatusList": {
      "post": {
        "operationId": "issueRevocationStatusList",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IRevocationListDataArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/IRevocationStatus"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/revokeCredential": {
      "post": {
        "operationId": "revokeCredential",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IHashCredentialArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialStatus"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
    "/activateCredential": {
      "post": {
        "operationId": "activateCredential",
        "description": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IHashCredentialArgs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialStatus"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Error name"
                    },
                    "method": {
                      "type": "string",
                      "description": "Method"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "path": {
                      "type": "string",
                      "description": "Error path"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    },
                    "description": {
                      "type": "string",
                      "description": "Error description"
                    }
                  },
                  "required": [
                    "didUrl"
                  ],
                  "description": "Agent method arguments validation error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "IDIDManagerFindArgs": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "description": "Optional. Alias"
          },
          "provider": {
            "type": "string",
            "description": "Optional. Provider"
          }
        },
        "description": "Input arguments for  {@link IDIDManager.didManagerFind | didManagerFind }"
      },
...
      "IIdentifier": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "description": "Decentralized identifier"
          },
          "alias": {
            "type": "string",
            "description": "Optional. Identifier alias. Can be used to reference an object in an external system"
          },
          "provider": {
            "type": "string",
            "description": "Identifier provider name"
          },
          "controllerKeyId": {
            "type": "string",
            "description": "Controller key id"
          },
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IKey"
            },
            "description": "Array of managed keys"
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IService"
            },
            "description": "Array of services"
          }
        },
        "required": [
          "did",
          "provider",
          "keys",
          "services"
        ],
        "description": "Identifier interface"
      },
      "IDIDManagerGetArgs": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "description": "DID"
          }
        },
        "required": [
          "did"
        ],
        "description": "Input arguments for  {@link IDIDManager.didManagerGet | didManagerGet }"
      },
      "IDIDManagerCreateArgs": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "description": "Optional. Identifier alias. Can be used to reference an object in an external system"
          },
          "provider": {
            "type": "string",
            "description": "Optional. Identifier provider"
          },
          "kms": {
            "type": "string",
            "description": "Optional. Key Management System"
          },
          "options": {
            "type": "object",
            "description": "Optional. Identifier provider specific options"
          }
        },
        "description": "Input arguments for  {@link IDIDManager.didManagerCreate | didManagerCreate }"
      },
...
      "IHashCredentialArgs": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string"
          }
        },
        "required": [
          "hash"
        ]
      },
...
    }
  }
}